+++ src/main/java/es/udc/fi/dc/irlab/rm/SumReducer.java
@@ -19,7 +19,7 @@ package es.udc.fi.dc.irlab.rm;
 import java.io.IOException;
 import java.util.Iterator;
 
-import org.apache.hadoop.io.FloatWritable;
+import org.apache.hadoop.io.DoubleWritable;
 import org.apache.hadoop.io.Writable;
 import org.apache.hadoop.mapreduce.Reducer;
 
@@ -27,20 +27,20 @@ import org.apache.hadoop.mapreduce.Reducer;
  * Emit <x, sum_y A_xy> from <x, {A_xy}>.
  */
 public class SumReducer extends
-	Reducer<Writable, FloatWritable, Writable, FloatWritable> {
+	Reducer<Writable, DoubleWritable, Writable, DoubleWritable> {
 
     @Override
-    protected void reduce(Writable key, Iterable<FloatWritable> values,
+    protected void reduce(Writable key, Iterable<DoubleWritable> values,
 	    Context context) throws IOException, InterruptedException {
 
-	Iterator<FloatWritable> it = values.iterator();
+	Iterator<DoubleWritable> it = values.iterator();
 
-	float sum = it.next().get();
+	double sum = it.next().get();
 	while (it.hasNext()) {
 	    sum += it.next().get();
 	}
 
-	context.write(key, new FloatWritable(sum));
+	context.write(key, new DoubleWritable(sum));
 
     }
